handle lack of uid and gid for devnode callback
The kernel patch adding uid and gid to the devnode callback for struct
device_type is defining DRIVER_CORE_DEVICE_TYPE_DEVNODE_UID, so trigger
on that to know if we should be setting these values or not.

This will go away in the final version, for now this means if you don't
have the kernel patch, everything is owned by root.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>