kdbus.h: turn data_ref into a real struct and rename it to kdbus_vec
"ref" suggests a relation to ref counting which this struct doesn't have
anything to do with. Instead this is much more like a (64bit-clean)
iovec, hence let's call it "vec" from now on.

This also gives the struct a proper name so that we can easily use
sizeof on it without having to cast a NULL instance.