kmsg: store a short-hand to the type of kernel-generated notification types
A kernel-generated notification mesaage will only contain one single
payload type, so instead of searching for all possible data attachments,
we can as well just refer to a stored copy from match.c.

As fds are never passed with along with such a message, we can store the
two fields in a union.