copy messages directly from sender into a receiver-supplied buffer
Require the receiver to pre-allocate a buffer and register it with
HELLO. The sender will copy its data directly into the revceiver's
buffer.

The RECV call will now only return a pointer to the next message
in the buffer. The buffer needs to be free()d with the FREE ioctl.

FIXME: currently missing is a proper allocator to manage the
receiver's buffer, in the current state the allocator will only clear
the buffer if all messages are free()d. This is obviously only useful
for testing.