names.c: simplify buffer size logic
Assume that clients send KDBUS_NAME_INFO_ITEM_NAME as their first
appended item, so we don't actually have to walk the items.

Also, do a put_user on the struct's total size, so the client knows
which extra size to assume. After all, that depends on
CONFIG_AUDITSYSCALL and CONFIG_SECURITY.