names: allow listing all names, including unique names
Introduce a flags field to struct kdbus_cmd_names, and
KDBUS_NAME_LIST_UNIQUE_NAMES to let the kernel know to not only dump
well-known-names but unique names as well.

That allows userspace to get a full list of all connected clients.

Not that user space implementation will need to adopt this change.