connection: implement new 'reply expected' logic
Instead of punching temporary holes into the policy database, maintain a
list of connections that a connection expects a reply from, along with a
timeout value and the cookie in place.

This makes the code easier to read, and allows us to check for missing
replies within the same timeout check that we already have.