connection: kick off timeout scan for conn_src, not conn_dst
conn_src is the one we added the reply list to, so this one should be
kicked off.

Note that the first invocation of the timeout scanner will unlikely do
any real work anyway. The reason for calling it here is that it will
implicitly calculate a new deadline, and arm the timer.