connection: let conn->reply_list keep track of allowed replies
Currently, when connection A sends a message to connection B, we store
the reply item with connection A, indicating that connection B is
allowed to reply to connection A.

This creates an ABBA locking issue with MSG_DROP, as we have to hold
both locks in order to do the list_del(reply->entry).

Let's do it the other way around and let conn->reply_list store the
reply items of allowed replies *from* the connection instead. That also
simplifies a number of all sites, as we don't have to walk all
connection on the bus anymore in order to find out whether anyone is
waiting for a reply from us.