names: fix return codes and conditions
The conn->names is decreased once the implementor takes over a name from
the activator. Hence, we can't rely on it to check whether a connection
already is the holder of an activatable name.

Refactor the code so that the error codes match:

-EALREADY is returned when the connection already is the owner of the
name is has request to become (activator or implementor)

-EINVAL is returned if the connection is an activator but already owns
a different name entry