policy: attach policy db to bus
The policy uploaded by a registrar connection must be stored with the
bus, not the endpoint. There will be a second instance of a policy db
for custom eps later, so keep the pointer around. Also, we'll need more
lookup logic for that as well.

Note that this is still work in progress.