connection: allow updating policy with kdbus_cmd_conn_update()
If the list of items passed along via the KDBUS_CMD_CONN_UPDATE ioctl
contains any KDBUS_ITEM_NAME or KDBUS_ITEM_POLICY_ACCESS item, the
policy database for the bus of the connection is updated. This will
happen atomically, in the way that all rules previously uploaded from
this connection will be removed, and the new set will be installed.