policy: make kdbus_policy_check_see_access() unlocked
We have a race condition in names, because kdbus_name_list_write() is
called twice - once for gathering the size of the resulting list, and
once for the actual write.

We need to make sure that the policy doesn't make different decisions in
both cases, so we have to hold the policy db's entries_lock around all
calls of kdbus_name_list_all().

For this, struct kdbus_policy_db has to be made public. Also, rename
kdbus_policy_check_see_access() to
kdbus_policy_check_see_access_unlocked().