metadata: kdbus_meta_append_auxgroups(): actually set the group ids
As we need metadata in other context than queue items, we don't want
to rely on the queue item logic to clean up after us and fix up the
values. For now, add back the code to actually fill the payload. We
might add code to make the queue patching conditional later.