metadata, queue: remove all namespace translating hacks again
After carefully revisiting the entire namespace translation ideas
again, we concluded that what we're currently doing doesn't possibly
cover all use cases that we have in mind.

a) We want to be able to do create a connection inside a container,
   pass the file handle to it to the outside world, and expect to be
   operating from inside the container with it. Namespace translation
   is not on in such cases.

b) While we can map pids and uids from inside a namespace into
   something that is meaningful on the host, the other way around is
   impossible: if we pass a handle from the host into the container,
   namespace translation cannot succeed.

As user namespaces aren't something we need to support anytime soon
anyway, the easiest solution for now is to remove all the current
code again and remove the metadata from the final message in case
we're jumping across namespaces with it.