util: disallow padding as part of strings in ioctl-commands
Up until now, we allow ioctl-commands with strings at the end to contain
padding after the string-terminator. We don't allow this in any other
item and we want "p->size - offsetof(typeof(*p), name)" to be the size of
the string +1.

By requiring the p->size field to be set to the exact string-length +1, we
avoid any compatibility issues and improperly sized fields.

sd-bus already does set p->size correctly, and everyone sane should do the
same.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>