handle: dont unref connection if copying the result fails
If we turned an handle into an connected endpoint, we must not unref the
connection if we cannot copy the result into user-space. The endpoint has
already been created and set on handle->conn. A following close() will
thus unref it again.

Instead of trying to revert the connection establishment, keep it but
still return EFAULT. If users supplied invalid buffers, there is something
seriously wrong, anyway. Therefore, expecting users to continue with that
handle is just insane. Thus, take the fast-path and simply keep the handle
in its new state.

Besides, this code currently fails, anyway. We have to disconnect the
conn-handle first, before we can drop the last reference.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>