handle: fix race on parallel type transformations
If two ioctls are called on a handle in parallel, and both try to turn
that handle into another type, we currently will fail horribly.

As there is no generic lock we can take across any handle transformation,
we avoid locking entirely and instead make sure a type transformation is
atomic. If we change a handle type, we lock the parent domain, update the
type and context and release the domain.

If another transformation was already done at that time, we bail out and
revert any newly created context.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>