match: make sure temporary entries are properly deleted
If we create a match-entry, but fail building up the rules list, we will
call kdbus_match_entry_free() on an unlinked match-entry. This will fail
horribly as it calls list_del() on a zeroed list entry.

Initialize the list entry via INIT_LIST_HEAD(). This will trick list_del()
into thinking the entry is linked, even though it's not.
This is a kinda ugly hack, but it's how we do list-entry protection in the
kernel (list_del() on an empty list-head is a no-op).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>