Revert "handle: drop redundant conn->type checks"
This reverts commit d229fb947f874cfa335a1c510729842f6a4db3f4.

These checks actually look for the connection type, not whether the
connection is disconnected. We really want to disallow most ioctls on
non-connection type FDs, so restore the old behavior.

However, we should really avoid storing the DISCONNECTED state in the
"type" field. If we extract it, we can fix the error-code mismatches that
the original commit tried to fix.