connection: introduce new API to check the connection types
As noted in commit 7729afbdc93e4d to properly report the type of a
connection we should store the DISCONNECTED state in its own variable
and not in the "type" field.

This was introduced in commit 7015a1e6746e0c to prevent special-purpose
connections from owning names and exchangin messages. However this
behaviour can be considered racy as noted by David in commit d229fb947f
and which was reverted.

Currently we may return different error types, consider if a connection
was a normal connected connection KDBUS_CONN_CONNECTED, then two CPUs
that share the same connection may hit:

CPU 1						CPU2

conn_a->type == KDBUS_CONN_CONNECTED		conn_a->type = KDBUS_CONN_CONNECTED
...						...
kdbus_conn_disconnect()				KDBUS_CMD_MSG_SEND ioctl()
=> conn_a->type == KDBUS_CONN_DISCONNECTED	...
...						if (conn_a->type != KDBUS_CONN_CONNECTED)
							return -EOPNOTSUPP;

This is a bug, since that ioctl() should not fail with -EOPNOTSUPP.
It should fail with -ECONNRESET but not at that level as it was also
noted by David, but deep down, when we are operating on the connection's
data and when we can safely take the locks.

So to fix this I am introducing some inline functions into headers to
check if the connection has the right type. It seems that the use of
inlined functions in header files is not popular in kdbus which can be
considered a good thing, since we can easily bloat the header files.
In the other hand if we are sure that GCC LTO for the kernel is working
well we can just move them into connection.c. Currently I don't know the
state of LTO on the kernel.

The important thing of using these API functions, it will allow to check
the connection type using the old methode before commit 7015a1e6746e0c.
We just use the "conn->flags" field which is set only during connection
creation time. This guarentees that we will have the correct connection
type without involving or taking any lock. Yes the connections once
created can't change their type.

So with this no need to add another connection field to separate the
state of the connection from its type, we will use the 'flags' field to
get the connection type, and convert the 'type' field to the old
'disconnected' one.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>
(fixed some typos)
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>