test: add race-testing
This adds some rather trivial helpers to test for cross-thread races when
we run multiple ioctls in parallel. If allows you to define code-blocks
that run before/after the test-run, and one code-block for each thread.
The threads are started in random order and synchronized via spin-locks,
so the start time is as close as possible.

Furthermore, we allow you to run tests multiple times and examine all the
results that occurred. Since we randomize thread execution, and you can
add random usleep()-calls, we can try to verify all possible races were
exercised. This is, obviously, not reliable, so we only return TEST_SKIP
in case something didn't go as expected.

The tests so far are rather trivial. More tests will follow.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>