notify: fix race between notify-flush and bus-disconnect
bus->ep is the default endpoint of a bus. As it has a reference to the bus
itself, we have to drop it when disconnecting a bus. However, flushing the
notification queues is done async (as it needs the bus lock) on bus->ep.
Make sure we own a reference during the flush.

Note that if the bus is disconnected while we currently flush the queue,
then kdbus_conn_kmsg_send() will notice it. In the upper layer, we just
need to make sure the endpoint stays accessible.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>