bus: add kdbus_bus_cred_is_privileged()
This is equivalent to kdbus_bus_uid_is_privileged() but takes a cred
pointer to test against the bus. We will need this for the updated
policy-db code to test conn->cred against a bus.

Note that capabilities are never taken from "conn->cred" but always from
current(). This has two reasons: for one, we want to be able to drop
capabilities from a task and *really* loose them. Secondly,
security_capable() (which is required to tests caps against creds) is not
exported (probably for a reason). So always use the current creds.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>