policy: make bus-owner checks use conn->cred, not current()
When doing policy checks, we always use the credentials associated with
the connection, not the ones of the thread. This is essential if we want
to allow processes to drop privileges but retain privileges via the
connection itself.

This is essentially the same we do for files via file->f_cred.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>