connection: fix race in kdbus_conn_new()
In kdbus_conn_new(), trigger to notification chain after all connection
details have been set up and its list entry is linked.

This fixes a race when subscribers are woken up and trying to address the
new connection but fail, because the connection is not yet linked up in
the list.

Signed-off-by: Daniel Mack <daniel@zonque.org>