connection: enforce message queue and reply count limits on CAP_IPC_OWNER
Limit the maximum number of message queues and pending replies by checking
for CAP_IPC_OWNER rather than whether the current user is 'privileged' by
the means of kdbus.

The problem with kdbus_bus_uid_is_privileged() is that it returns true if
the current user is the bus owner, and so on user buses, connection could
queue arbitrary amounts of data.

Signed-off-by: Daniel Mack <daniel@zonque.org>