connection: sleep on sender queues during sync calls
Instead of a wait-queue for each message, we now use the sender's
wait-queue to wait for replies. This queue is woken up on each incoming
reply, so it seems reasonable to use it for sync calls, too.

This way, we avoid adding wait-queues to reply-calls that are inaccessible
from kdbus_conn_disconnect() (at least in reasonable time). By using the
conn->wait queue, we can easily synchronize sync-send-calls and
disconnects.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>