handle: acquire connections for all ioctls
If we work on connections, we really want to avoid those connections to
get disconnected while we work on it. All our ioctls correctly wake up if
a connection is shut down, so lets acquire connections for all ioctls.
This guarantees that kdbus_conn_disconnect() now blocks until all ioctl
contexts finished.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>