pool slice: introduce slice->public
When a user calls KDBUS_CMD_MSG_RECV with KDBUS_RECV_PEEK, we return the
offset of a message in the pool and allow the user to call KDBUS_CMD_FREE
on it. That, however, is a problem as the message remains in the pool at
this point, and later, when the user calls KDBUS_CMD_MSG_RECV again, we
access invalid memory.

Fix this by introducing a 'public' flag in a slice, and check for it
when KDBUS_CMD_FREE is called. Only if the slice was flagged public
before, allow the user to free it. Otherwise, return -EINVAL. The flag
is set by normal message delivery, connection info and name listing, but
not of the message was just PEEKed or if the user guesses an offset.

Signed-off-by: Daniel Mack <daniel@zonque.org>