endpoint: make sure EP_UPDATE can clear policies
There is no reason to disallow items_size==0 for EP policy updates. If an
endpoint should not have a single policy, let them achieve that easily.
Note that they could just install a policy for a random name instead, to
achieve no effective policy is set. So allow items_size==0 to avoid hacks
in user-space.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>