domain: embed 'struct device'
We currently allocate a dynamic "struct device" object as domain->dev.
There is really no need to do that. We can safely embed it as part of
"struct kdbus_domain" and get rid of our own ref-counter. Instead, use the
ref-counter of the device.

This way, container_of(device, struct kdbus_domain, dev) on "struct
device" will return a pointer to our kdbus-domain. This will be very handy
once we embed char-devs, too. We can finally get rid of our own IDRs and
use the char-dev IDRs for lookup.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>