kdbus.h: Use #defines rather than an enum for ioctl definition
Arnd Bergmann writes:

I think in general, using enum is great, but for ioctl command numbers,
we probably want to have defines so the user space implementation can
use #ifdef to see if the kernel version that it is being built for
knows a particular command.

[...]

which might in fact help some tools that try to do automated parsing
of header files to find ioctl commands.

Signed-off-by: Daniel Mack <daniel@zonque.org>