handle.c: rework pointer assignment logic
During the review on LKML, Thomas Gleixner stubled over the usage
of our 'p' variable assignment. Apparantly, the idea of assigning
memdup()'ed memory to a void pointer that is automatically freed
at the end of the function wasn't obvious to readers.

Let's fix this, and

a) make kdbus_memdup_user() return void* instead of int, so we can
   directly assign variables to the return value

b) assign the void* variable after the memdup, and call it 'free_ptr'
   to make clearer what's going on.

Signed-off-by: Daniel Mack <daniel@zonque.org>