metadata: rework attach logic
In queue.c, we used to append the entire metadata object to each
message, and ignored the fact that we might attach more data to a
message that the receiving peer has originally requested.

Change that now, and introduce two new functions in metadata.c:

  kdbus_meta_size()
  kdbus_meta_write()

They are both given a mask of attach items, and one will just report
the size that excerpt is about to take, the other will write that
excerpt to a kdbus_pool_slice.

That buys as a number of things:

  a) we don't copy more data to the receiving peer than it requested

  b) we can now implement peer-to-peer matrix specific attach logic,
     and drop items on behalf of the sender or the receiver

  c) it's easier to implement namespace translation logic and just
     translate or drop some specific items per receiver

Signed-off-by: Daniel Mack <daniel@zonque.org>